attempt to fix search params vercel build issue
