+++ app/pages/beta-login/page.tsx
@@ -1,77 +1,13 @@
-"use client";
+// app/pages/confirm-email/page.tsx
+// Server Component wrapper for custom email confirmation
 
-import { useState } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-
-export default function BetaLoginPage() {
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const redirect = searchParams.get("redirect") || "/";
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setError(null);
-    setLoading(true);
-
-    try {
-      const res = await fetch("/api/beta-login", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ password }),
-      });
-
-      const data = await res.json();
-      if (!res.ok) {
-        setError(data.error || "Invalid password");
-        return;
-      }
-
-      router.push(redirect);
-      router.refresh();
-    } catch {
-      setError("Something went wrong. Please try again.");
-    } finally {
-      setLoading(false);
-    }
-  }
-
+import { ClientPageWrapper } from '@/components/auth/client-page-wrapper'
+import BetaLoginPage from '@/components/auth/betaloginpage'
+export default function ConfirmEmailPage() {
   return (
-    <div className="min-h-screen flex items-center justify-center bg-background">
-      <Card className="w-full max-w-sm">
-        <CardHeader>
-          <CardTitle className="text-lg font-semibold">
-           This site is in beta testing phase. Enter access password
-          </CardTitle>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="space-y-2">
-              <label className="text-sm font-medium">Password</label>
-              <Input
-                type="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                autoFocus
-              />
-            </div>
-            {error && (
-              <p className="text-sm text-destructive">
-                {error}
-              </p>
-            )}
-            <Button type="submit" className="w-full" disabled={loading}>
-              {loading ? "Checking..." : "Unlock"}
-            </Button>
-          </form>
-        </CardContent>
-      </Card>
-    </div>
-  );
+    <ClientPageWrapper>
+      <BetaLoginPage />
+    </ClientPageWrapper>
+  )
 }
+
